fraud_detection_backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.4
django-cache-machine==1.2.0
djangorestframework==3.15.1
joblib==1.3.2
kombu==5.3.6
numpy==1.26.4
pandas==2.2.1
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.4.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
